#623 Storage refactor
