+++ WMIAdventure/frontend/src/js/storage/cache/cache.js
