+++ WMIAdventure/frontend/src/js/storage/localStorageKeys.js
